TW第三周作业 扩展需求，提出新的优惠：  1.购买某种商品满100减一定数额（最低优先级的优惠）  2.总花费满100减一定数额  3.解析器抽象  4.calculate接口

+++ POS/src/main/java/com/wrongkey/pos/model/RegularItem.java
@@ -9,20 +9,20 @@ package com.wrongkey.pos.model;
  */
 public class RegularItem extends Item {
 
-    public RegularItem(String barcode,float price,int quantity) {
+    public RegularItem(String barcode,int price,int quantity) {
         this.barcode = barcode;
         this.price = price;
         this.quantity = quantity;
     }
 
     /**
-     * @return float
+     * @return int
      * @author wrongkey
      * @description calculate the cost
      * @date 2014/11/20
      */
     @Override
-    public float calculateTheCost() {
+    public int calculateTheCost() {
         return getPrice()*getQuantity();
     }
